Refactor towhee.ops to support pipeline

Update repo uri


Remove repo manager


Update for pylint


Mv operators/pipelines to hub
