+++ towhee/engine/factory.py
@@ -13,9 +13,14 @@
 # limitations under the License.
 import os
 import threading
-from typing import Tuple
+from typing import List, Tuple
 
+from towhee.dataframe import DataFrame
+from towhee.pipeline_format import OutputFormat
+from towhee.engine.pipeline import Pipeline
+from towhee.engine.engine import Engine, start_engine
 from towhee.engine.operator_loader import OperatorLoader
+from towhee.hub.file_manager import FileManager
 from towhee.hparam.hyperparameter import CallTracer, param_scope
 # pylint: disable=unused-import
 from towhee.hub import preclude
@@ -141,17 +146,159 @@ class _OperatorLazyWrapper:
         with param_scope(index=self._index) as hp:
             self._op = op(self._name, self._tag, **self._kws)
 
+DEFAULT_PIPELINES = {
+    'image-embedding': 'towhee/image-embedding-resnet50',
+    'image-encoding': 'towhee/image-embedding-resnet50',  # TODO: add encoders
+    'music-embedding': 'towhee/music-embedding-vggish',
+    'music-encoding': 'towhee/music-embedding-clmr',  # TODO: clmr -> encoder
+}
+
+class _PipelineWrapper:
+    """
+    A wrapper class around `Pipeline`.
+
+    The class prevents users from having to create `DataFrame` instances by hand.
+
+    Args:
+        pipeline (`towhee.Pipeline`):
+            Base `Pipeline` instance for which this object will provide a wrapper for.
+    """
+    def __init__(self, pipeline_: Pipeline):
+        self._pipeline = pipeline_
+
+    def __call__(self, *args) -> List[Tuple]:
+        """
+        Wraps the input arguments around a `Dataframe` for Pipeline.__call__(). For
+        example:
+        ```
+        >>> p = pipeline('some-pipeline')
+        >>> result = p(arg0, arg1)
+        ```
+        """
+        if not args:
+            raise RuntimeError('Input data is empty')
+
+        cols = []
+        vargs = []
+        for i, arg in enumerate(args):
+            vtype = type(arg).__name__
+            cols.append(('Col_' + str(i), str(vtype)))
+            vargs.append(arg)
+        vargs = tuple(vargs)
+
+        # Process the data through the pipeline.
+        in_df = DataFrame('_in_df', cols)
+        in_df.put(vargs)
+        out_df = self._pipeline(in_df)
+        format_handler = OutputFormat.get_format_handler(self._pipeline.pipeline_type)
+        return format_handler(out_df)
+
+    def __repr__(self) -> str:
+        return repr(self._pipeline)
+
+    @property
+    def pipeline(self) -> Pipeline:
+        return self._pipeline
+
+
+def pipeline(pipeline_src: str, tag: str = 'main', install_reqs: bool = True):
+    """
+    Entry method which takes either an input task or path to an operator YAML.
+
+    A `Pipeline` object is created (based on said task) and subsequently added to the
+    existing `Engine`.
+
+    Args:
+        pipeline_src (`str`):
+            pipeline name or YAML file location to use.
+        tag (`str`):
+            Which tag to use for operators/pipelines on hub, defaults to `main`.
+        install_reqs (`bool`):
+            Whether to download the python packages if a requirements.txt file is included in the repo.
+
+    Returns
+        (`typing.Any`)
+            The `Pipeline` output.
+    """
+    start_engine()
+
+    if os.path.isfile(pipeline_src):
+        yaml_path = pipeline_src
+    else:
+        fm = FileManager()
+        p_repo = DEFAULT_PIPELINES.get(pipeline_src, pipeline_src)
+        yaml_path = fm.get_pipeline(p_repo, tag, install_reqs)
+
+    engine = Engine()
+    pipeline_ = Pipeline(str(yaml_path))
+    with param_scope() as hp:
+        if not hp().towhee.dry_run(False):
+            engine.add_pipeline(pipeline_)
+
+    return _PipelineWrapper(pipeline_)
+
+
+class _PipelineBuilder:
+    """
+    Build a pipeline with template variables.
+
+    A pipeline template is a yaml file contains `template variables`,
+    which will be replaced by `variable values` when createing pipeline instance.
+
+    Examples:
+    ```yaml
+    name: template_name
+    variables:                           <<-- define variables and default values
+        template_variable_1: default_value_1
+        template_variable_2: default_value_2
+    ....
+
+    operator:
+        function: {template_variable_1}  <<-- refer to the variable by name
+    ```
+
+    You can specialize template variable values with the following code:
+
+    >>> pipe = _PipelineBuilder(template_variable_1='new_value').pipeline('pipeline_name')
+    """
+    def __init__(self, **kws) -> None:
+        self._kws = kws
+
+    def pipeline(self, *arg, **kws):
+        with param_scope() as hp:
+            hp().variables = self._kws
+            return pipeline(*arg, **kws)
+
+    @staticmethod
+    def callback(name, index, *arg, **kws):
+        name = name.replace('.', '/').replace('_', '-')
+        _ = index
+        return _PipelineBuilder(**kws).pipeline(name, *arg)
+
+
+def _ops_call_back(real_name: str, index: Tuple[str], *arg, **kws):
+    try:
+        return _PipelineBuilder.callback(real_name, index, *arg, **kws)
+    except: # pylint: disable=bare-except
+        return _OperatorLazyWrapper.callback(real_name, index, *arg, **kws)
+
 
 class OpsCallTracer(CallTracer):
     """
     Entry point for creating operator instances, for example:
 
-    >>> op_instance = ops.my_namespace.my_operator_name(init_arg1=xxx, init_arg2=xxx)
+    >>> op_instance = ops.my_namespace.my_repo_name(init_arg1=xxx, init_arg2=xxx)
 
-    An instance of `my_namespace`/`my_operator_name` is created.
+    An instance of `my_namespace`/`my_repo_name` is created. It will automatically judge if the repo is an operator or a pipeline in hub,
+    but it will take a little time, you can also specify the `rtype`('pipeline' or 'operator') manually, for example:
+
+    >>> op_instance = ops.my_namespace.my_op_name(init_arg1=xxx, init_arg2=xxx, rtype='operator')
+    >>> pipe_instance = ops.my_namespace.my_pipe_name(init_arg1=xxx, init_arg2=xxx, rtype='pipeline')
+
+
     """
     def __init__(self, callback=None, path=None, index=None):
         super().__init__(callback=callback, path=path, index=index)
 
 
-ops = OpsCallTracer(_OperatorLazyWrapper.callback)
+ops = OpsCallTracer(_ops_call_back)
